+++ src/views/MainView.vue
@@ -19,6 +19,7 @@
                 <li class="nav-item">
                     <a class="nav-link" :class="{active: (activeTab === 'selection')}" href="#" @click="activeTab = 'selection'">
                       <font-awesome-icon icon="shopping-cart"></font-awesome-icon> Cart
+                      <span class="badge badge-secondary">{{selectedVariableIds}}</span>
                     </a>
                 </li>
             </ul>
@@ -42,7 +43,7 @@ import ContentView from './ContentView.vue'
 import SidebarView from './SidebarView.vue'
 import CartView from './CartView.vue'
 import ToastComponent from '../components/ToastComponent.vue'
-import { mapState, mapMutations, mapActions } from 'vuex'
+import { mapState, mapGetters, mapMutations, mapActions } from 'vuex'
 
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faStore, faShoppingCart } from '@fortawesome/free-solid-svg-icons'
@@ -60,7 +61,11 @@ export default Vue.extend({
     }
   },
   computed: {
-    ...mapState(['toast', 'isSignedIn'])
+    ...mapState(['toast']),
+    ...mapGetters(['isSignedIn']),
+    selectedVariableIds () {
+      return Object.keys(this.$store.state.gridSelection).length
+    }
   },
   methods: {
     ...mapMutations(['clearToast', 'setToast']),