+++ tests/unit/store/actions.spec.ts
@@ -1,5 +1,4 @@
 import actions from '@/store/actions'
-import router from '@/router'
 import { Cart } from '@/types/Cart'
 import emptyState from '@/store/state'
 import orders from '../fixtures/orders'
@@ -196,10 +195,6 @@ jest.mock('@molgenis/molgenis-api-client', () => {
     delete_: function () { mockDelete(...arguments) }
   }
 })
-
-jest.mock('@/router', () => ({
-  push: jest.fn()
-}))
 
 describe('actions', () => {
   describe('loadOrders', () => {
@@ -546,6 +541,35 @@ describe('actions', () => {
         post.mockResolvedValue('success')
         await actions.save({ state, commit })
         expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order', expect.anything(), true)
+        expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Saved order with order number 12345' })
+        done()
+      })
+    })
+
+    describe('if applicationForm is a fileRef', () => {
+      it('saves order', async (done) => {
+        const commit = jest.fn()
+        const state: ApplicationState = {
+          ...emptyState,
+          order: {
+            orderNumber: '12345',
+            name: null,
+            projectNumber: null,
+            applicationForm: {
+              id: 'id',
+              url: 'url',
+              filename: 'my file'
+            },
+            state: null,
+            submissionDate: null,
+            creationDate: null,
+            updateDate: null
+          }
+        }
+        jest.spyOn(orderService, 'buildFormData').mockImplementation(() => new FormData())
+        post.mockResolvedValue('success')
+        await actions.save({ state, commit })
+        expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order/12345?_method=PUT', expect.anything(), true)
         expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Saved order with order number 12345' })
         done()
       })