+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+# [0.13.0](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.12.0...v0.13.0) (2019-11-19)
+
+
+### Bug Fixes
+
+* **main:** don't chain the calls ([d3898d9](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/d3898d9))
+* import umd instead of sources ([07efd35](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/07efd35))
+
+
+### Features
+
+* show version in footer ([8a936fd](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/8a936fd))
+
+# [0.12.0](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.11.0...v0.12.0) (2019-11-19)
+
+
+### Features
+
+* add selected variables badge to Cart tab ([0e47789](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/0e47789)), closes [#150](https://github.com/molgenis/molgenis-app-lifelines-webshop/issues/150)
+
+# [0.11.0](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.7...v0.11.0) (2019-11-18)
+
+
+### Bug Fixes
+
+* vscode eslint validation & autofix ([54279b1](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/54279b1))
+
+
+### Features
+
+* add ConfirmationModal component for order delete ([ec2a6f3](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/ec2a6f3))
+* ConfirmationModal unit & integration test ([a90354c](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/a90354c))
+
+## [0.10.7](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.6...v0.10.7) (2019-11-14)
+
+
+### Bug Fixes
+
+* Update order with file data ([4224cde](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/4224cde))
+
 ## [0.10.6](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.5...v0.10.6) (2019-11-13)